ModelBay_BackEnd/requirements.txt
flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
huggingface-hub==0.34.4
requests==2.31.0
